Update workflow with mistletoe rendering and make clean up changes
Updated the exploration script to work with mistletoe for markdown rendering to ensure improved compatibility. Mistletoe, a python markdown parser, is used to replace md2notion methods for markdown file parsing. This is accomplished by creating a new NotionPyRenderer class in the Renderer.py script where different rendering functions are defined based on mistletoe's BaseRenderer.

A new requirements.txt file was also created to reflect this new dependency on the mistletoe library.

This commit also includes code clean up changes in the journal file and the exploration script for better readability and to amend spacing inconsistencies. Additionally, a new configuration option was added in the .idea/misc.xml file to allow for early release upgrades for python packages.